Fixed timestep physics

+++ src/evan/game/gamestate/GameState.java
@@ -13,7 +13,7 @@ public abstract class GameState {
 	
 	public abstract void init();
 	
-	public abstract void update(float delta);
+	public abstract void update(double delta);
 	
 	public abstract void render(Graphics g);
