+++ src/evan/game/gamestate/MenuState.java
@@ -22,7 +22,7 @@ public class MenuState extends GameState {
 	}
 
 	@Override
-	public void update(float delta) {
+	public void update(double delta) {
 		
 	}
